requirements.txt
google-cloud-aiplatform==1.35.0
google-cloud-documentai==2.20.2
python-dotenv==1.0.0